fs: Statfs_t{} doesn't have a Type field on NetBSD, OpenBSD, or Solaris
Other projects that import procfs, like node_exporter, apparently build
for OSs that don't have Statfs_t.Type.
